+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+ENHANCEMENTS
+
+* Add notifications for Security Hub findings via SNS topic LandingZone-SecurityHubFindings ([#56](https://github.com/schubergphilis/terraform-aws-mcaf-landing-zone/pull/56))
+
 ## 0.4.5 (2021-01-06)
 
 BUG FIXES
@@ -23,6 +29,10 @@ BUG FIXES
 ENHANCEMENTS
 
 * Set default password policy parameters for the audit, logging, master accounts ([#57](https://github.com/schubergphilis/terraform-aws-mcaf-landing-zone/pull/57))
+
+ENHANCEMENTS
+
+* Forward SecurityHub findings to AggregateSecurityNotifications SNS topic ([#56](https://github.com/schubergphilis/terraform-aws-mcaf-landing-zone/pull/56))
 
 ## 0.4.2 (2020-12-29)